move fixtures back to test/ from test/browser for sharing with server-side tests
