+++ package.json
@@ -5,7 +5,7 @@
   "main": "build/index.js",
   "scripts": {
     "build": "babel lib -d build",
-    "build-test": "webpack --config test/browser/fixtures/webpack.config.js",
+    "build-test": "webpack --config test/fixtures/webpack.config.js",
     "docs": "rm -rf docs && jsdoc lib -t node_modules/minami -R README-API.md -d docs",
     "prepublish": "npm run build",
     "test": "npm run build-test && npm run test-server && npm run test-browser-local",
