+++ test/browser/index.html
@@ -4,7 +4,7 @@
     <meta charset="utf-8">
     <script src="../../node_modules/webcomponents.js/webcomponents.js"></script>
     <script src="../../node_modules/web-component-tester/browser.js"></script>
-    <script src="fixtures/build.js"></script>
+    <script src="../fixtures/build.js"></script>
   </head>
   <body>
     <script src="test-component.js"></script>
